Add saveAsFile function to export data as Excel file
Added a new script function `saveAsFile` to the `App.razor` file.
This function allows the application to export data as an Excel
file. It creates a Blob from a byte array with the MIME type for
Excel files, generates a download link, and programmatically
clicks the link to trigger the download.